feat: node controller
-Add component to edit sensor and device node settings via sidebar menu.
