+++ client/src/components/context/SensorDeviceContext.tsx
@@ -10,6 +10,7 @@ export const SensorDeviceProvider = ({
 }) => {
   const [sensors, setSensors] = useState<Sensor[]>([]);
   const [devices, setDevices] = useState<Device[]>([]);
+  const [selectedNodeId, setSelectedNodeId] = useState<string | null>(null);
 
   // Optional: Init user if needed
   useLocalStorage({ actionType: "user-init-update" });
@@ -19,6 +20,7 @@ export const SensorDeviceProvider = ({
     actionType: "init",
     setSensors,
     setDevices,
+    setSelectedNodeId,
   });
 
   // Sync whenever sensors/devices update
@@ -26,10 +28,14 @@ export const SensorDeviceProvider = ({
     actionType: "sync",
     sensors,
     devices,
+    selectedNodeId,
+    setSelectedNodeId,
   });
 
   return (
-    <SensorContext.Provider value={{ sensors, setSensors }}>
+    <SensorContext.Provider
+      value={{ sensors, setSensors, selectedNodeId, setSelectedNodeId }}
+    >
       <DeviceContext.Provider value={{ devices, setDevices }}>
         {children}
       </DeviceContext.Provider>
