+++ client/src/components/context/SensorDeviceContextDefs.ts
@@ -4,6 +4,8 @@ import type { Device, Sensor } from "../utils/other/Types";
 export interface SensorContextType {
   setSensors: React.Dispatch<React.SetStateAction<Sensor[]>>;
   sensors: Sensor[];
+  selectedNodeId: string | null;
+  setSelectedNodeId: React.Dispatch<React.SetStateAction<string | null>>;
 }
 
 export const SensorContext = createContext<SensorContextType | null>(null);
