+++ client/src/pages/Home.tsx
@@ -11,10 +11,10 @@ export default function Home() {
           <div className="flex absolute left-1/2 -top-1/3 transform -translate-x-1/2  text-slate-400">
             Floor 4
           </div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
         </Link>
       </div>
 
@@ -26,10 +26,10 @@ export default function Home() {
           <div className="flex absolute left-1/2 -top-1/3 transform -translate-x-1/2  text-slate-400">
             Floor 3
           </div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
         </Link>
       </div>
       <div className="flex flex-col gap-8">
@@ -40,10 +40,10 @@ export default function Home() {
           <div className="flex absolute left-1/2 -top-1/3 transform -translate-x-1/2  text-slate-400">
             Floor 2
           </div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
         </Link>
       </div>
       <div className="flex flex-col gap-8">
@@ -54,10 +54,10 @@ export default function Home() {
           <div className="flex absolute left-1/2 -top-1/3 transform -translate-x-1/2  text-slate-400">
             Floor 1
           </div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
-          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md h-full"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
+          <div className="group-hover:bg-slate-400 flex border-2 border-slate-500 w-full h-full min-w-40 min-h-25 rounded-md"></div>
         </Link>
       </div>
     </div>
